+++ common/spectrum.c
@@ -12,15 +12,15 @@
 
 #endif
 
-double spectrum_sum(Spectrum *s)
+Float spectrum_sum(Spectrum *s)
 {
-    double output = 0;
+    Float output = 0;
     for(int i = 0; i < SPECTRUM_SIZE; i++)
         output += *s[i];
     return output;
 }
 
-double spectrum_avg(Spectrum *s)
+Float spectrum_avg(Spectrum *s)
 {
     return spectrum_sum(s) / SPECTRUM_SIZE;
 }
