+++ include/transform.h
@@ -19,9 +19,9 @@ Transform transform_combine(Transform t1, Transform t2);
 Transform transform_translate(vec3 delta);
 Transform transform_scale(vec3 scale);
 
-Transform transform_rotate_x(double angle);
-Transform transform_rotate_y(double angle);
-Transform transform_rotate_z(double angle);
+Transform transform_rotate_x(Float angle);
+Transform transform_rotate_y(Float angle);
+Transform transform_rotate_z(Float angle);
 
 vec3 transform_apply_vector(vec3 p, Transform t);
 vec3 transform_apply_point(vec3 p, Transform t);
